requirements.txt
confluent-kafka
websocket-client
pyspark==3.5.0